% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_activeyr.R
\name{citec_activeyr}
\alias{citec_activeyr}
\title{Returns the year the author started publishing}
\usage{
citec_activeyr(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A number indicating the author started activity.
}
\description{
Returns the year the author started publishing
}
