% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_affil.R
\name{citec_affil}
\alias{citec_affil}
\title{Returns the affiliation of the author}
\usage{
citec_affil(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A string indicating institution the author is affiliated with.
}
\description{
Returns the affiliation of the author
}
