% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_allids.R
\name{citec_allids}
\alias{citec_allids}
\title{Scrapes RePEc website to find all ids}
\usage{
citec_allids()
}
\value{
A vector of all ids on https://ideas.repec.org/i/eall.html.
Alternatively you can find ids based on names
here https://ideas.repec.org/cgi-bin/shortid.cgi
}
\description{
Scrapes RePEc website to find all ids
}
