% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_authors_df.R
\name{citec_authors_df}
\alias{citec_authors_df}
\title{Returns a dataframe of authors' information for a given vector of author ids.}
\usage{
citec_authors_df(ids)
}
\arguments{
\item{ids}{A vector of RePEc ids. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A dataframe containing author's id, name, affiliation, number of papers, number of
citations, and years of being active.
}
\description{
Returns a dataframe of authors' information for a given vector of author ids.
}
