% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_df.R
\name{citec_df}
\alias{citec_df}
\title{Returns a dataframe of all the publications of a given vector of ids
including information on author's name, affiliations, number of papers
and citations, etc.}
\usage{
citec_df(ids)
}
\arguments{
\item{ids}{A vector of RePEc ids. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A dataframe containing the title of each publication, its type,
its publication year, the number of citations received, and the publishing
journal. For each row, it also contains information on the author of the
publicaiton including name, affiliation, number of papers, number of
citations, and years of being active.
}
\description{
Returns a dataframe of all the publications of a given vector of ids
including information on author's name, affiliations, number of papers
and citations, etc.
}
