% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_hic.R
\name{citec_hic}
\alias{citec_hic}
\title{Returns the h-index, the i10-index, and the number of citations of the author}
\usage{
citec_hic(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A list containing the h-index, the i10-index, and the number of citation for each author
}
\description{
Returns the h-index, the i10-index, and the number of citations of the author
}
