% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_name.R
\name{citec_name}
\alias{citec_name}
\title{Returns the name of the author}
\usage{
citec_name(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A string indicating the name of the author
}
\description{
Returns the name of the author
}
