% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_otherpeople.R
\name{citec_otherpeople}
\alias{citec_otherpeople}
\title{Returns people who frequently work with the author,
people who cite the author frequently and people who
are cieted by the author frequently.}
\usage{
citec_otherpeople(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A list with three elements: vector of people who work
with the author, vector of people who are cited by the author, vector of
people who have cited the author
}
\description{
Returns people who frequently work with the author,
people who cite the author frequently and people who
are cieted by the author frequently.
}
