% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_output.R
\name{citec_output}
\alias{citec_output}
\title{Returns the productivity measures of the author}
\usage{
citec_output(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A list with two elements: the number of articles and
the number of papers published by an author
}
\description{
Returns the productivity measures of the author
}
