% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_paperlist.R
\name{citec_paperlist}
\alias{citec_paperlist}
\title{Returns information on the publications published by the author}
\usage{
citec_paperlist(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A dataframe containing the title of each publication, its type,
its publicaiton year, and the number of citations received.
}
\description{
Returns information on the publications published by the author
}
