% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_papers_df.R
\name{citec_papers_df}
\alias{citec_papers_df}
\title{Returns a dataframe of all the publications of a given vector of ids
including author's name.}
\usage{
citec_papers_df(ids)
}
\arguments{
\item{ids}{A vector of RePEc ids. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A dataframe containing the title of each publication, its type,
its publication year, the number of citations received, the publishing
journal and the author of the publicaiton.
}
\description{
Returns a dataframe of all the publications of a given vector of ids
including author's name.
}
