% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citec_selfcite.R
\name{citec_selfcite}
\alias{citec_selfcite}
\title{Returns the number of self-cites by the author}
\usage{
citec_selfcite(id)
}
\arguments{
\item{id}{The RePEc id of the author. RePEc id for
each author can be found here: https://ideas.repec.org/cgi-bin/shortid.cgi}
}
\value{
A number indicating the number of times an author has cited themselves
}
\description{
Returns the number of self-cites by the author
}
